+++ src/group.rs
@@ -26,6 +26,46 @@ pub struct Scroll {
     _inner: *mut Fl_Scroll,
 }
 
+impl Scroll {
+    /// Returns the x position
+    pub fn xposition(&self) -> u32 {
+        unsafe {
+            Fl_Scroll_xposition(self._inner) as u32
+        }
+    }
+
+    /// Returns the y position
+    pub fn yposition(&self) -> u32 {
+        unsafe {
+            Fl_Scroll_yposition(self._inner) as u32
+        }
+    }
+
+    /// Scrolls from ```from``` to ```to```
+    pub fn scroll_to( &mut self, from: u32, to: u32) {
+        debug_assert!(from <= std::i32::MAX as u32, "u32 entries have to be < std::i32::MAX for compatibility!");
+        debug_assert!(to <= std::i32::MAX as u32, "u32 entries have to be < std::i32::MAX for compatibility!");
+        unsafe {
+            Fl_Scroll_scroll_to(self._inner, from as i32, to as i32)
+        }
+    }
+
+    /// Gets the scrollbar size
+    pub fn scrollbar_size(&self) -> u32 {
+        unsafe {
+            Fl_Scroll_scrollbar_size(self._inner) as u32
+        }
+    }
+
+    /// Sets the scrollbar size
+    pub fn set_scrollbar_size(&mut self, new_size: u32) {
+        debug_assert!(new_size <= std::i32::MAX as u32, "u32 entries have to be < std::i32::MAX for compatibility!");
+        unsafe {
+            Fl_Scroll_set_scrollbar_size(self._inner, new_size as i32)
+        }
+    }
+}
+
 /// Creates a tab which can contain widgets
 #[derive(WidgetExt, GroupExt, Debug)]
 pub struct Tabs {